requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
groq==0.11.0
httpx==0.27.0
pydantic==2.9.0
python-multipart==0.0.12

# NLP & Sentiment Analysis
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2
transformers==4.36.2

# Vector Database & RAG
faiss-cpu==1.7.4
numpy==1.24.3

# Data Processing
pandas==2.0.3
scikit-learn==1.3.1

# Audio Transcription
assemblyai==0.17.0

# Environment
python-dotenv==1.0.0
